+++ src/Entity.cpp
@@ -1,9 +1,9 @@
 #include "Entity.h"
 
-void Entity::updateVx(int amount) {
+void Entity::updateVx(float amount) {
     vx = vx + amount;
 }
 
-void Entity::updateVy(int amount) {
+void Entity::updateVy(float amount) {
     vy = vy + amount;
 }
